requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# HTTP Client for Async Scraping
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.2

# AI & PDF Processing
groq==0.4.2
pypdf==4.0.1

# Database & Environment
supabase==2.3.7
python-dotenv==1.0.1

# Optional but Recommended for Resume Parsing
python-docx==1.1.0